refactor: move common functionalities out of Ops.cpp
