+++ src/enzyme_ad/jax/Dialect/Utils.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include "mlir/Interfaces/SideEffectInterfaces.h"
+#include "llvm/ADT/SmallVector.h"
+
+namespace mlir {
+namespace enzymexla {
+
+void addMemoryEffectsFromAttr(
+    SmallVectorImpl<MemoryEffects::EffectInstance> &effects,
+    ArrayAttr effectsAttr);
+
+void addAllMemoryEffects(
+    SmallVectorImpl<MemoryEffects::EffectInstance> &effects);
+
+} // namespace enzymexla
+} // namespace mlir
